refactor(chatbi): AskUserDialog 移到 chat/components 目录
